backend/requirements-mac.txt
flask
pandas
numpy
scikit-learn
yfinance
tensorflow-macos
tensorflow-metal
textblob
python-dotenv
flask-cors